+++ packages/wiz-ui-react/src/components/custom/index.ts
@@ -1,3 +1,4 @@
+export * from "./chat";
 export * from "./form";
 export * from "./notification";
 export * from "./schedule";